requirements.txt
#langChain Core
langChain
langchain-Core
httpcore
httpx

#OpenAI Integration
langchain-openai
openai

#Anthropic intergration
langchain-anthropic

#Google gemini (PaLM) Integration
langchain-google-genai
google-generativeai

#Hugging face Integration
langchain-huggingface
transformers
huggingface-hub

#Environment Variable management
python-dotenv

#Machine Leearning Utilities
numpy
scikit-learn
matplotlib
ipykernel
torch
torchvision